Update README.md
How to Create a WordPress Child Theme?
Create a theme folder. It is recommended to suffix the directory name with “-child”. For example, if the parent theme name is ‘foobar’, it is recommended to create the folder name as “foobar-child”.
Create style.css in this folder. It should have a standard header comment as given below.
Create functions.php
That’s it! you are ready with your minimal WordPress Child theme. Child theme should be put under theme directory and activated as usual.